[fix] AuthForm
Make LoginForm and Signup form for increase bugs.
When we create login form and signup form with together, we got problem while register validation.
So I ignore the forename and surname in mapLoop array but, the button submit not working and i think that because we dont declare or don't set value correct value for forename and surname fields.